Fix CI: clang 19/20, vulkan=true, add linting, fix clang-tidy warnings
Agent-Logs-Url: https://github.com/jwinarske/drm-cxx/sessions/cc6446fa-e4b1-4aa7-94d3-1c4927dadf6c

+++ drm-cxx/planes/plane_registry.hpp
@@ -16,7 +16,7 @@ class Device;
 
 namespace drm::planes {
 
-enum class DRMPlaneType : uint32_t {
+enum class DRMPlaneType : uint8_t {
   PRIMARY,
   OVERLAY,
   CURSOR,
